pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "openaedmap-backend"
readme = "README.md"
version = "0.0.0"

[tool.poetry.dependencies]
alembic = "^1.13.1"
anyio = "^4.2.0"
asyncache = "<1"
authlib = "^1.3.0"
beautifulsoup4 = {extras = ["charset-normalizer", "lxml"], version = "^4.12.3"}
cachetools = "^5.3.2"
fastapi = "<1"
feedgen = "^1.0.0"
httpx = {extras = ["brotli", "http2"], version = "<1"}
jinja2 = "^3.1.2"
mapbox-vector-tile = "^2.0.1"
msgspec = "<1"
numpy = "^1.26.3"
pillow = "^10.2.0"
pyinstrument = "^4.6.1"
pyproj = "^3.6.1"
python = "^3.12"
python-magic = "<1"
python-multipart = "<1"
pytz = "*"
scikit-learn = "^1.3.2"
sentry-sdk = {extras = ["fastapi", "httpx", "sqlalchemy"], version = "^1.44.0"}
shapely = "^2.0.2"
sqlalchemy = {extras = ["postgresql-asyncpg"], version = "^2.0.29"}
supervisor = "^4.2.5"
tzfpy = "*"
uvicorn = {extras = ["standard"], version = "<1"}
xmltodict = "<1"
zstandard = {extras = ["cffi"], version = "<1"}

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]

indent-width = 4
line-length = 120
target-version = "py312"

[tool.ruff.lint]
ignore = [
  "ARG001", # unused argument
  "ARG002", # unused positional argument
]
select = [
  # see https://docs.astral.sh/ruff/rules/ for rules documentation
  "A", # flake8-builtins
  "ARG", # flake8-unused-arguments
  "ASYNC", # flake8-async
  "B", # flake8-bugbear
  # "COM", # flake8-commas
  "C4", # flake8-comprehensions
  "E4", # pycodestyle
  "E7",
  "E9",
  "F", # pyflakes
  # "FBT", # flake8-boolean-trap
  "FLY", # flynt
  # "FURB", # refurb (preview)
  "G", # flake8-logging-format
  "I", # isort
  "INT", # flake8-gettext
  # "LOG", # flake8-logging (preview)
  "N", # pep8-naming
  "NPY", # numpy
  "UP", # pyupgrade
  "Q", # flake8-quotes
  "PERF", # perflint
  "PGH", # pygrep-hooks
  "PIE", # flake8-pie
  # "PL", # pylint
  "PT", # flake8-pytest-style
  "PTH", # flake8-use-pathlib
  "PYI", # flake8-pyi
  "RSE", # flake8-raise
  "RUF", # ruff
  "S", # flake8-bandit
  "SIM", # flake8-simplify
  "SLF", # flake8-self
  "SLOT", # flake8-slots
  "T10", # flake8-debugger
  # "T20", # flake8-print
  # "TRY", # tryceratops
  "YTT", # flake8-2020
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "lf"

[tool.ruff.lint.flake8-builtins]
builtins-ignorelist = ["id", "open", "type", "filter", "format"]

[tool.ruff.lint.flake8-gettext]
extend-function-names = ["pgettext", "npgettext"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "double"

[tool.ruff.lint.pep8-naming]
extend-ignore-names = ["ET"]

[tool.ruff.lint.pylint]
max-args = 10
